pyproject.toml
[tool.poetry]
name = "interview-assistant"
version = "0.1.0"
description = "AI-powered interview question generator for technical roles"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "interview_assistant", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.1"
openai = "^1.3.7"
python-docx = "^1.1.0"
PyPDF2 = "^3.0.1"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
numpy = "^1.24.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.3.0"

[tool.poetry.scripts]
interview-assistant = "interview_assistant.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 